+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.30.1 (TBD)
 
 - Fixed rare null deref caused by mastery ritual code querying GameManager without a valid GameManager instance
+- Added mastery descriptions to [ItemTips](https://enter-the-gungeon.thunderstore.io/package/Glorfindel/ItemTips/?utm_source=discord) output
 
 ## 1.30.0 (2026-05-07)